requirements.txt
# EnzyParser Dependencies
docling>=2.64.0
pydantic>=2.0
pandas>=2.0
pillow>=10.0
openpyxl>=3.1
google-generativeai>=0.5.0
python-dotenv>=1.0.0
